# Replace regexp parser with normal parser of directive and arguments